chip and footer tests finished
